+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## In Progress ##
 
-- **Added:** functional.caller
+- **Added:** functional.caller, functional.suppress
 - **Changed:** retry.retry_if_errors. Used `functools.wraps` to apply update_wrapper.
 - **Changed:** file.decompress. Added `size` arg.
 - **Changed:** Purpose of `n` arg of algorithm.fibonacci.